+++ src/components/Posts/ExpandedPostItem.tsx
@@ -33,7 +33,7 @@ import { Link as Routerlink } from "react-router-dom";
 import UserAvatar from "../Common/UserAvatar";
 import { useTheme } from "@mui/material/styles";
 import PostMenu from "./PostMenu";
-import TooltipTimestamp from "../../utilities/ToolTipTimestamp";
+import TooltipTimestamp from "../Common/ToolTipTimestamp";
 
 const styles = {
   actionButton: {
